new features: line modification and lines bookmarks
now each block has a private block (or not!) which states it status. the status is used
to display the line background, markers on the lines panel, and set the modification status 
of each line.

if a line is modified a green light will be displayed. this can be used to known what did you
modify since you loaded that file.