Fix a bug on replace
When you open first the replace widget, and search for a non
existing text, when you closed the replace widget there was no
cursor on the text widget.

This was because by default m_searchCursor is a NULL cursor, and it
needs to have a real value when you close the replace widget.